+++ lib/core/init.js
@@ -9,14 +9,11 @@ import {
 	waitForChild,
 	waitForEvent,
 	newSitetable,
-	markStart,
-	markEnd,
 } from '../utils';
-import * as Modules from './modules';
 import { _loadModuleOptions } from './options/options';
-import { _loadModulePrefs } from './modules/modules';
+import { _loadModulePrefs, _runModuleStage } from './modules/modules';
+import { _addModuleBodyClasses } from './modules/bodyClasses';
 import { migrate } from './migrate';
-import type { BooleanOption, EnumOption } from './module';
 
 let start;
 
@@ -69,7 +66,7 @@ export const runMigrations: Promise<void> = loadI18n
 	.then(() => migrate());
 
 export const loadDynamicOptions: Promise<void> = loadI18n
-	.then(() => allModules('loadDynamicOptions', { skipEnabledCheck: true }));
+	.then(() => _runModuleStage('loadDynamicOptions', { skipEnabledCheck: true }));
 
 export const loadOptions: Promise<*> = loadDynamicOptions
 	.then(() => Promise.all([
@@ -81,73 +78,18 @@ export const addModuleBodyClasses: Promise<void> = loadOptions
 	.then(() => _addModuleBodyClasses());
 
 export const always: Promise<void> = loadOptions
-	.then(() => allModules('always', { skipEnabledCheck: true }));
+	.then(() => _runModuleStage('always', { skipEnabledCheck: true }));
 
 export const beforeLoad: Promise<void> = loadOptions
-	.then(() => allModules('beforeLoad'));
+	.then(() => _runModuleStage('beforeLoad'));
 
 export const go: Promise<*> = Promise.all([beforeLoad, bodyReady])
 	.then(() => {
 		initObservers();
-		return Promise.all([newSitetable(document.body), allModules('go')]);
+		return Promise.all([newSitetable(document.body), _runModuleStage('go')]);
 	});
 
 export const afterLoad: Promise<void> = Promise.all([go, loadComplete])
-	.then(() => allModules('afterLoad'));
-
-const ERRORED_KEY = Symbol('errored');
-
-async function allModules(key, { skipEnabledCheck = false }: {| skipEnabledCheck?: boolean |} = {}) {
-	await Promise.all(
-		Modules.all()
-			.filter(module => (
-				module[key] &&
-				!module[ERRORED_KEY] &&
-				(skipEnabledCheck || Modules.isRunning(module))
-			))
-			.map(async module => {
-				const tag = markStart();
-				try {
-					const stage = module[key];
-					await stage();
-				} catch (e) {
-					module[ERRORED_KEY] = true;
-					console.error('Error in module:', module.moduleID, 'during:', key);
-					console.error(e);
-				}
-				markEnd(tag, `${module.moduleID} (${key})`);
-			})
-	);
-}
-
-// Adds body classes for modules or enabled options that have `bodyClass: true`
-// In the form `res-moduleId-optionKey` for boolean options
-// and `res-moduleId-optionKey-optionValue` for enum options
-// spaces in enum option values will be replaced with underscores
-function _addModuleBodyClasses() {
-	for (const module of Modules.all()) {
-		if (!Modules.isRunning(module)) continue;
-
-		if (module.bodyClass) BodyClasses.add(`res-${module.moduleID}`);
-
-		for (const [optId, opt] of Object.entries(module.options)) {
-			if (!(opt.bodyClass && opt.value)) continue;
-			// ensure only boolean and enum options can pass this check
-			(opt: BooleanOption<any> | EnumOption<any>);
-
-			if (opt.dependsOn && !opt.dependsOn(module.options)) continue;
-
-			let cls = typeof opt.bodyClass === 'string' ?
-				opt.bodyClass :
-				`res-${module.moduleID}-${optId}`;
-
-			if (opt.type === 'enum') {
-				cls += `-${opt.value.replace(/\s/g, '_')}`;
-			}
-
-			BodyClasses.add(cls);
-		}
-	}
-}
+	.then(() => _runModuleStage('afterLoad'));
 
 bodyStart.then(() => BodyClasses.add());
