+++ lib/core/modules/modules.js
@@ -5,6 +5,8 @@ import { Storage, i18n } from '../../environment';
 import {
 	downcast,
 	matchesPageLocation,
+	markStart,
+	markEnd,
 } from '../../utils';
 import { Module } from '../module';
 import type { OpaqueModuleId } from '../module';
@@ -41,6 +43,30 @@ export async function _loadModulePrefs() {
 			enabled[id] = !modules[id].disabledByDefault;
 		}
 	}
+}
+
+const ERRORED_KEY = Symbol('errored');
+export async function _runModuleStage(stage: $Keys<Module<any>>, { skipEnabledCheck = false }: {| skipEnabledCheck?: boolean |} = {}) {
+	await Promise.all(
+		all()
+			.filter(module => (
+				module[stage] &&
+				!module[ERRORED_KEY] &&
+				(skipEnabledCheck || isRunning(module))
+			))
+			.map(async module => {
+				const tag = markStart();
+				try {
+					const fn = module[stage];
+					await fn();
+				} catch (e) {
+					module[ERRORED_KEY] = true;
+					console.error('Error in module:', module.moduleID, 'during:', stage);
+					console.error(e);
+				}
+				markEnd(tag, `${module.moduleID} (${stage})`);
+			})
+	);
 }
 
 export function setEnabled(opaqueId: OpaqueModuleId, enable: boolean) {
